refactor(labware-creator): Move conditional labware diagrams to own file
closes #7706
